moved all components into folders, changed all css to scss
