+++ src/components/LoadingAnimation/LoadingAnimation.scss
