+++ src/components/ServicesContainer/ServicesContainer.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
-import Service from './Service/Service';
-import './ServicesContainer.css';
+import Service from '../Service/Service';
+import './ServicesContainer.scss';
 
 function ServicesContainer() {
   const servicesData = useSelector((state) => state.servicesReducer.servicesData);
