+++ src/components/SubmittingAnimation/SubmittingAnimation.scss
