+++ src/pages/FaqPage/FaqPage.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import QuestionsContainer from '../../components/QuestionsContainer';
-import './FaqPage.css';
+import QuestionsContainer from '../../components/QuestionsContainer/QuestionsContainer';
+import './FaqPage.scss';
 
 function FaqPage() {
   return (
